requirements.txt
peewee==3.16.0
python-dotenv==1.0.0
PyJWT~=2.6.0
Flask~=2.2.3